Add order shipping rate endpoint
This adds the initial boilerplate for the new shipping rates endpoint.
We're adding the route to the existing orders controller. The
permissions check is the same check that happens for a single order. The
only difference here is that we're adding a response handler
specifically for the shipping methods. It's not implemented right now,
but we need to calculate the cost for all available shipping methods for
the current order.

Closes #30794